requirements.txt
python-dotenv
python-dateutil
PyMySQL
mfrc522
RPi-GPIO
RPLCD
smbus2